gui-libs/eglexternalplatform: add 1.2
The new meson.build changes the header install location to
/usr/include rather than /usr/include/EGL but should be
harmless given the old .pc file was adding -I/usr/include/EGL
and egl-gbm/wayland are not using EGL/ in #include statements.

Currently unclear if this version is safe to use with <nvidia-560,
I believe it only adds access to a new function and it shouldn't
matter if it is not used (yet). Did give it a quick try but may
have been insufficient to expose issues.

Upstream commit message:
  With version 1.2 (supported starting in 560 series drivers), platform
  libraries can safely call into the driver during
  unloadEGLExternalPlatform. Prior to that, the driver would have torn
  down the internal EGLDisplays by that point.

Ultimately 560 is due to replace 550 as the stable driver, and old
versions bare 470 (w/o USE=wayland anyway) are due for eventual
removal. So *eventual* usage may be a non-issue.

If need to check for driver version on this package, may need a subslot
to force rebuilds when driver version changes, hopefully can avoid.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>