sci-chemistry/pymol: drop a bunch of dependencies to runtime-only
They aren't needed to run setup.py (whether via gpep517 or otherwise).
They are only used by the runtime code. The distinction between
dependency types is sometimes seen as a bit academic, but it matters for
--buildpkgonly. This is especially useful as one of the runtime-only
dependencies will, in turn, depend on python[tk], and moving it out of
DEPEND means the package can be test-compiled with --buildpkgonly,
without rebuilding and reinstalling python.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>