linux-mod-r1.eclass: BDEPEND on dev-util/pahole
If the kernel is built with CONFIG_DEBUG_INFO_BTF_MODULES, external
kernel module builds will try to call `pahole` to generate BTF and
fail if it's not present.

pahole is a small dependency and we're likely to enable BTF by default
in more configurations going forward, so do it unconditionally.

The alternatives aren't great anyway, we would have to instead:
* RDEPEND in g-k-bin
* BDEPEND in g-k-bin and break the binpkg case
* Use an ewarn in the eclass if pahole isn't installed + config option is set
so this is the best option available to us.

Thanks-to: Ionen Wolkens <ionen@gentoo.org>
Reported-by: Sebastian Engel <sighunter@gmx.de>
Signed-off-by: Sam James <sam@gentoo.org>