net-vpn/wireguard-modules: drop ~ia64
We're about to drop ~ia64 support entirely anyway but also this pkg
is only useful for specific kernel versions. It's likely the intersection
of people on such kernels & ia64 is the empty set.

(Need to do this because of linux-mod-r1 adding a dep on pahole in
82f9bef4ed76f656a3d79b607eecff526b78b7ee and this is a better option
than bothering to test pahole on ia64, sorry.)

Signed-off-by: Sam James <sam@gentoo.org>