kde-frameworks/kirigami: backport System Settings cache fix
Was planning on pulling this in as a workaround for now given it's
a nasty bug but it's been merged upstream now too. Not doing
straight-to-stable as might introduce performance regressions so let's
give it a (small) amount of soaking time first.

KDE-bug: https://bugs.kde.org/show_bug.cgi?id=488326
Signed-off-by: Sam James <sam@gentoo.org>