profiles/targets/desktop: make USE=qml default for more than just plasma
Installing random applications tends to drag in qml requirements, as it
is a pivotal part of the Qt technology stack required by other
components. One quickly ends up in USE flag resolution hell when trying
to select it on a per-package basis for most use cases other than having
Qt installed solely for an isolated application.

For average desktop use it makes sense to simply default it to on.
People can always disable it manually if they want, but it basically
just adds a single package (qtdeclarative) as a dep to your Qt stack.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>