dev-libs/librdkafka: mark as LTO-unsafe
Upstream has an innovative bug report strategy. Wait for people to
report bugs in the software, then ignore the bug report for several
years, then show up and as the very first official reply:

> Closing due to inactivity, does not seem like a common issue.

We can presume that this will never be fixed. Although the nuance of
whether it is broken depends on the exact architecture used, it is not
worth figuring out where it is safe to use for such a, well,
*kafkaesque* bit of software. :)

Bug: https://github.com/confluentinc/librdkafka/issues/2426
Closes: https://bugs.gentoo.org/895464
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>