x11-drivers/nvidia-drivers: use PreserveVideoMemoryAllocations=1
(Disclaimer: I do not have the right setup to test any of this, but
doing it blind given the increasing amount of affected users -- call
for testing has shown that it should work as expected albeit users
with more custom setups such as suspend without elogind/systemd will
need to pay attention to warnings given breakage *is* expected).

Was formerly disabled because it broke sleep with elogind, systemd if
the units were not enabled, and other custom methods to enable sleep.

However, =0 is limited and is seemingly broken with wayland (typically
resulting in graphical corruption after resume). GDM straight up refuse
to show a wayland session if it's not set wrt bug #873160, and several
Plasma 6 users been reporting issues and its upstream also instructs
distributions to set this.

So this adds a elogind hook, enables systemd units by default (much
like it is for the elogind hook), and at least warn for the last case
which is considered semi-unsupported.

elogind does have its own HandleNvidiaSleep option, but it is intended
for old drivers which did not ship a nvidia-sleep.sh and reports seem to
show that it may not be working properly. Ebuild warns that it should be
disabled instead, and also tries to warn if there is old custom scripts
installed by the user.

One downside of hook vs the option is that hooks are not told if using
suspend or hibernate and this sends the wrong message to the drivers
(albeit not known to be an issue at the moment).

May not fix everything wrt bug #693384, but believe this is the best
we can do downstream unless someone knows better, and so closing it.
There are plenty of issues unrelated to elogind too, ideally would
need users to compare with systemd before filing more elogind sleep
bugs unless know exactly what is causing issues in elogind.

wrt bug #873160, this only fix *one* thing that the gdm udev rules
check and so may not mean will necessarily start seeing wayland in
gdm. Rules currently need =1, systemd-only, and a non-hybrid setup
(aka just nvidia, no offloading). See also the general bug #939201.

Straight-to-stable may not be the best idea, but wanted to simplify
and not revbump the 3 .conf, duplicate them, and adjust every ebuilds
further for this (believe it *should* be ok, or at least not make
things worse for typical users). Also want to deliver the fix early
to plasma 6 users newly using wayland by default.

Closes: https://bugs.gentoo.org/693384
Closes: https://bugs.gentoo.org/873160
Closes: https://github.com/gentoo/gentoo/pull/38482
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>