x11-drivers/nvidia-drivers: use PreserveVideoMemoryAllocations=1
(Disclaimer: I do not have the right setup to test any of this, but
doing it blind w/ feedback given the increasing amount of affected
users).

Was formerly disabled because it broke sleep with elogind, systemd if
the units were not enabled, and other custom methods to enable sleep.

However, =0 is limited and is seemingly broken with wayland and some
DEs. GDM straight up refuse to show a wayland session if it's not set
wrt bug #873160, and several Plasma 6 users been reporting issues
(upstream also instructs distributions to set this).

So this adds a elogind hook, force enabling units by default (much like
it is for the elogind hook), and at least warn for the last case which
is considered semi-unsupported.

elogind does have its own HandleNvidia option, but it is intended for
old drivers which did not ship a nvidia-sleep.sh and reports show that
it may not be working properly. Ebuild warns that it should be disabled
instead. One downside of hook vs the option is that hooks aren't told
if using suspend or hibernate and this sends the wrong message to the
drivers (albeit not known to be an issue at the moment).

May not fix everything wrt bug #693384, but believe this is the best we
can do downstream unless someone knows better.

Straight-to-stable may not be the best idea, but wanted to simplify
and not revbump the 3 .conf, duplicate them, and adjust every ebuilds
further for this (believe it *should* be ok, or at least not make
things worse).

Closes: https://bugs.gentoo.org/693384
Closes: https://bugs.gentoo.org/873160
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>