net-libs/libssh2: pass build option to disable static-libs
And apply patch to make the (anemic) testsuite run under this scenario.

Previously, we dropped USE=static-libs when switching to build with
cmake, then upstream brought support for that back and defaulted it to
on -- but we built it and deleted the raw file instead, which broke
cmake-config since that does runtime file checks to verify whether all
originally installed files that you aren't using, exist.

Since we tell the build system not to build+install static libraries at
all, the resulting cmake configs won't check for them!

Bug: https://bugs.gentoo.org/911386
Closes: https://bugs.gentoo.org/911548
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>