sci-libs/vtk: don't cap expat version
This reverts commit 546ead3103af153d5e869730fe0ad642019a9c67 and
d767aab3f077c05916d47ffdc38a09927b04298d as a consequence.

As I explained on the bug:
"""
This really isn't going to work either way.

If Python depends on >= expat to avoid things like bug 939045, you
soon get to a point where VTK is uninstallable if VTK has a < dep.

If it doesn't, you get runtime breakage with VTK if it has a < dep.
If VTK doesn't have a < dep, you get runtime breakage in Python as it
gets downgraded when you emerge VTK.

We should drop the < and leave VTK broken as it's the lesser evil,
and then try pull in the patch if it applies cleanlyish.
"""

So, let's do the lesser evil bit now to avoid runtime breakge
in CPython from it being downgraded.

Bug: https://bugs.gentoo.org/930032
Closes: https://bugs.gentoo.org/939211
Signed-off-by: Sam James <sam@gentoo.org>