app-emulation/wine-vanilla: update live (ffmpeg)
Haven't looked in-depth at how it's used and if it can currently serve
as a gstreamer replacement but, if so, may make sense to swap the
default eventually (both can be enabled too, but may just be wasteful).

Leaving default-off for now either way given it's new, and users that
weren't using multilib gst-plugins-libav already will need extra
multilib libraries.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>