app-emulation/wine-staging: update live (ffmpeg)
Haven't looked in-depth at how it's used and if it can currently serve
as a gstreamer replacement but, if so, may make sense to swap the
default eventually (both can be enabled too, but may just be wasteful).

Leaving default-off for now either way given it's new, and users that
weren't using multilib gst-plugins-libav already will need extra
multilib libraries.

(note: early update to match wine-vanilla-9999, technically wine-staging
has not rebased to have this change "yet" but at worst it'll just cause
this to give an unrecognized option for at most a few days if not hours)

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>