dev-util/libabigail: drop unconditional eautoreconf
We still need this for the live ebuild for obvious reasons. But we were
running it against release tarballs as well, with a comment from 2019
that said ltmain.sh contained broken redhat spec file references. This
is not a problem today (cannot find the relevant string, builds fine) so
it appears the release tooling they used is no longer broken and we can
remove this workaround.

All we need now is the usual elibtoolize dance, a much smaller
dependency set.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>