dev-debug/dtrace: rework pkg_postinst
For first installs, we enable the service and start it.

This is unusual, but the behaviour without dtprobed running is
untested/unsupported. It's not a network service, it has no configuration,
reads a single device node, and does all parsing within a seccomp jail. It
also leads to hard-to-diagnose issues because USDT probes won't be registered
and an application might have already started up which needs to be traced.

Signed-off-by: Sam James <sam@gentoo.org>