net-misc/r8152: add 2.18.1, update metadata
This introduces the next version of r8152, but also includes a change in
source. It is no longer possible to download the r8152 sources directly
from realtek without passing a captcha challenge. So, I imported the
tarballs directly into a new github project, documented the source and
the SHA256, and am now using releases from that as the source.

Signed-off-by: Jay Faulkner <jay@jvf.cc>
Closes: https://github.com/gentoo/gentoo/pull/38573
[Eli: dropped stable keyword for bump]
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>