sci-libs/mpfi: add 1.5.4_p20220401
Since the tarball names are different, we can call this a "patch
release" having otherwise the same version components.

The test suite is part of the tarball now, and of course, it fails. I
emailed the authors about this but haven't heard back yet. I think it's
more a case of the test suite being unfinished than of there being real
problems in the code. I've done my best to patch up the (3) failers.

Closes: https://bugs.gentoo.org/921351
Closes: https://github.com/gentoo/gentoo/pull/36024
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>