x11-drivers/nvidia-drivers: include nvidia.conf in dracut always
Dracut ignores /etc/modprobe.d if hostonly=no, this is a problem because if
plymouth is installed then the drm module will be pulled in which will include
the nouveau module. The presence of this module, combined with the lack of
modprobe.d config file blacklisting it, results in the nouveau module being
loaded in the initramfs stage on systems that have plymouth installed and are
not using dracut's hostonly setting.

This needs a proper upstream fix, but for now we can workaround the problem
with this.

See-also: https://github.com/dracut-ng/dracut-ng/issues/674
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>