dev-qt/qtwebengine: update patchsets in live (qt6)
6.7-12 -> 6.7-13, and 6.8-3 -> 6.8-4 changes:
  Updated:
  * cstdint.patch (merged extra chunk from files/)
  Dropped:
  * clang19.patch (upstreamed)
  * gcc15.patch (upstreamed)

Note:
6.8 patches are known to not apply right now, but that is because
they are made for the 6.8.0 branch (in preparation for release) while
the ebuild uses dev 6.8 which moved on to 6.8.1 and broke a patch.

Will update again sometime only after release is out of the way given
do not care to test 6.8.1 right now. Or earlier if the change is
backported to 6.8.0 in the interim.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>