x11-drivers/nvidia-drivers: config check for DRM_TTM_HELPER w/ 6.11+
Essentially the same issue as DRM_KMS_HELPER where nvidia will
detect that it's usable but the kernel will not install the
drm_ttm_helper.ko module resulting in missing symbol dependencies.

Also has the same problem where it cannot be directly selected
so we need to try to explain this to users.

At same time also update the KMS_HELPER message a bit, esp. given
I915 is not usable to enable TTM_HELPER making it a poor suggestion.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>