net-im/synapse: add legacy-cgi rdep for py3.13
Synapse uses cgi module, which was removed from py3.13. The package is
py3.13 enabled and everything worked because legacy-cgi runtime
dependency was indirectly installed as a dependency of treq. The treq
was recently release with official py3.13 support and it no longer needs
the legacy-cgi, which triggered the error with missing cgi module.
The synapse needs to depend on legacy-cgi directly anyway.

Signed-off-by: Petr Vaněk <arkamar@gentoo.org>