dev-lang/python: Do not remove ensurepip w/ USE=-ensurepip (3.12+)
Change the behavior of USE=-ensurepip in the new patchset bumps
not to remove the ensurepip package, but only control the dependency.
The flag was originally introduced to remove the package entirely
along with its bundled wheels.  However, since then we unbundled
the wheels, and the removal makes little sense — it's better to keep
it functional, so the wheels can be installed independently.

This change will be backported to older versions as we update
the patchsets there as well.

Signed-off-by: Michał Górny <mgorny@gentoo.org>