sys-libs/libstdc++-v3: disable custom user flags entirely
This is ancient brittle code -- an ancient, unmaintained version of GCC.
It exists solely to provide access to an ancient, unmaintained version
of libstdc++.so for ancient, unmaintained binaries that cannot be
rebuilt.

It doesn't need to be fast, or have cool CFLAGS. People shouldn't even
be using it at all. The people who do use it care about compatibility
more than anything else, so successfully compiling it should be taken as
the gift that it is.

With that in mind, we basically strip all flags. We preserve some
existing logic to handle -march= values known to exist in this old gcc,
and we continue to support -O levels, but we go no further. It's
entirely impractical to support arbitrary flags, since every gcc release
adds more flags that old versions don't support and then this fails
because GCC 3's internal xgcc build cannot handle them.

We also take this opportunity to add a workaround for modern C. New
versions of GCC barf on some of the code here because it is not legal
C99, but again, the point of this package is to provide legacy
compatibility for ancient binaries, so... we assume if it ever worked
before, it still works. :) Thus, suppress the errors. In order to
suppress the errors, we make a wrapper script around the external
compiler (not the xgcc one) so that we can apply flags that only affect
modern compilers, and aren't parsed by xgcc which would choke on them
due to those flags being too new.

Closes: https://bugs.gentoo.org/879775
Closes: https://bugs.gentoo.org/919184
Closes: https://bugs.gentoo.org/832016
Closes: https://bugs.gentoo.org/923112
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>