sys-boot/gnu-efi: Fix compile when LLVM `binutils-plugin` USE flag is set
LLVM's objcopy is being used instead of the GNU version when `binutils-plugin`
is enabled. I tried to avoid this by explictly using the GNU path for objcopy
when our initial test fails.

Let me know if this isn't acceptable, or if you need me to open a new bug
for this. It was basically the same repo as the linked issue, but the fix did
not work, as `which objcopy` returns '/usr/lib/llvm/18/bin/objcopy' instead of
'/usr/bin/objcopy'.

Bug: https://bugs.gentoo.org/931792
Signed-off-by: Maide <github@maide.ca>
Closes: https://github.com/gentoo/gentoo/pull/38785
Signed-off-by: Viorel Munteanu <ceamac@gentoo.org>