net-print/cups: backport PPD validation fixes
These fixes are in net-print/cups itself, which was not actually
referenced in the 4 CVEs in the CUPS writeup mentioned in bug 940312.

But they're also the only patches available right now, and they're clearly
related, so let's pull them in as others are doing too.

Specifically, it pulls in the following from the 2.4.x branch:
* 313c388dbc023bbcb75d1efed800d0cfc992a6cc
* 9939a70b750edd9d05270060cc5cf62ca98cfbe5
* 04bb2af4521b56c1699a2c2431c56c05a7102e69
* e0630cd18f76340d302000f2bf6516e99602b844
* 1e6ca5913eceee906038bc04cc7ccfbe2923bdfd
* 2abe1ba8a66864aa82cd9836b37e57103b8e1a3b

Bug: https://bugs.gentoo.org/940312
Bug: https://bugs.gentoo.org/940311
Bug: https://bugs.gentoo.org/940313
Bug: https://bugs.gentoo.org/940314
Bug: https://bugs.gentoo.org/940315
Bug: https://bugs.gentoo.org/940316
Signed-off-by: Sam James <sam@gentoo.org>