net-analyzer/yersinia: add 0.8.2_p20221119, drop 0.8.2-r2
current 0.8.2-r2 USE=gtk didn't appear to function correctly.  I fixed
the deps to be a bit more explicit but that didn't fix it.  Upgrading to
current git snapshot fixed the issue so I am removing the non-working
version and replacing it with a working version in one commit.

Additionally removing unneeded patches, thanks to Sam for upstreaming
them.
Signed-off-by: Rick Farina <zerochaos@gentoo.org>