profiles: unmask GCC 11 for now
GCC 11 is the last version of GDC which doesn't need GDC to build, so
it's useful for bootstrapping. Given the recent toolchain.eclass work,
it'd be a shame to not facilitate things working Out Of The Box until
we have some alternative package set up (see bug #940470).

So, unmask for now as a temporary reprieve. But really, please don't
use GCC 11 unless you must, as it's EOL.

This reverts the following commits:
* 4bb56f21a477a1cb4998bbc283be138bff483a9b
* 0f5f8e27a03786959c2f151ce6381ffc4200a9e7
* 7de3c29b71c2b46a1e1581dcc24525433982d687

(Really, a lot of the value in masking old GCCs is to push people
to check `gcc-config` and so on anyway, and most people will have
hopefully done that by now in response to the mask a week ago.

That doesn't make the mask pointless -- if anyone files a bug with GCC 11,
we're going to tell them to try >= GCC 12, and package maintainers aren't
expected to try make things work with < GCC 12 and so on, but it does
mean unmasking doesn't do much harm.)

Bug: https://bugs.gentoo.org/940470
Signed-off-by: Sam James <sam@gentoo.org>