www-client/chromium: `die` if rust profiler not found
Chromium's build system explicitly looks for
`libprofiler_builtins-*.rlib` when using a prebuilt / system Rust.

While we could patch this out so that users don't _need_
USE=profiler if not doing PGO, the Rust upstream default is to ship
the profiler; conditionally changing the Chromium build logic to
account for an untested and unsupported build is not desirable.

Instead we will rely on existing dependency checks
(which ensure that a Rust slot with the profiler is installed)
and do a sanity check on the selected Rust implementation before
unpacking Chromium's enormous tarball so that we can fail early and
prompt the user to `eselect` the correct version / slot.

Closes: https://bugs.gentoo.org/937161
Signed-off-by: Matt Jolly <kangie@gentoo.org>