sys-apps/portage: backport emerge-webrsync fix
As I said in the revert commit message, just pick the safe option for now
until I can be sure about the proper fix, as emerge-webrsync has been
brittle in the past.

Note that the important thing is that verification works and is on still,
i.e. the commit being reverted doesn't disable that.

Bug: https://bugs.gentoo.org/940120
Signed-off-by: Sam James <sam@gentoo.org>