gui-libs/egl-wayland: add 1.1.17_pre20240924
For testing explicit sync, has a lot of fixes over 1.1.16 but
release is taking a while and it makes it difficult to know if
issues are in egl-wayland or individual software when testing
meanwhile.

>=1.1.16 is still masked either way, odds are will remain that
way until nvidia-drivers ships with >=1.1.16 (currently doesn't)
and 1.1.17 is released.

Can drop src_install given it ships its own icd json file now.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>