app-emulation/wine-proton: add 9999, drop 9.0.9999
Using the "bleeding-edge" branch. Pretty sure this branch did not
exist when formerly added the live ebuild and so had settled for
the versioned experimental branch.

Do not wish to maintain two live ebuilds, and it probably make more
sense to use bleeding edge if any (experimental is not updated that
often, and may lack fixes for very recent games that some users may
need).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>