kde-plasma/drkonqi: enable py3.13, drop sentry-sdk dep
Per consultation with upstream, sentry-sdk is only used for reporting issues
in the GDB Python script (gdb_preamble/preamble.py). This is pretty niche
and not worth it given the issues mentioned in bug #937896.

drkonqi does rely otherwise on sentry but it implements the protocol/format
itself and not using the Python package.

This unblocks py3.13 too.

Bug: https://bugs.gentoo.org/937896
Signed-off-by: Sam James <sam@gentoo.org>