profiles: mask =dev-qt/*-6.8.0* and related for now
6.7.3 was just released either way w/ fixes (...and regressions that
are still present in 6.8.0 anyway) and should be no rush to get users
on 6.8.0.

Furthermore, better to get more users to keep using 6.7.3 for testing
given that is the stable candidate after some pending known regressions
been patched. 6.8.x is "likely" to wait until 6.8.1 for stable.

Typically been patching PyQt6 to allow building with the newer Qt, but
this time around it's non-trivial and we decided it'd be better off
masked until PyQt6-6.8.0 is released.

PySide6 has similar issues but that one is annoyingly pinned to minor
versions and wouldn't want to mask each of them. Thankfully less
packages depend on it, so it has less of an impact. Should "hopefully"
be updated too by the time of unmask though so pre-emptively add it to
the mask (ignoring UnknownProfilePackage pkgcheck warnings).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>