media-libs/fontconfig: fix aliasing violation
We had a report of GCC 14 with -O3 -flto causing wrong font
sizes with fontconfig (showing up in qalculate-gtk).

It turns out to be because _FcPatternIter and _FcPatternPrivateIter
are punned between which violates strict-aliasing rules, which
manifested in FcDefaultSubstitute getting a bogus value from
FcValueCanonicalize for size.

Mark the bad struct as may_alias to fix that.

Closes: https://bugs.gentoo.org/940923
Signed-off-by: Sam James <sam@gentoo.org>