www-client/chromium: make rust profiler sanity check more reliable
On ppc64le the machine hardware name printed out by 'uname -m'
is 'ppc64le' but the rust target is called 'powerpc64le'.
As a result, the build process will always die even if
we're using a rust compiler built with USE='profiler'.

Use this opportunity to tighten up some language in the ebuild
and elaborate more in comments.

Closes: https://bugs.gentoo.org/940729
Co-authored-by: zin0 <z_i_n_o@bluewin.ch>
Signed-off-by: Matt Jolly <kangie@gentoo.org>