app-emulation/libvirt: Record virtualbox dependency
The aim of virtualbox use flag is to drag in
app-emulation/virtualbox as dependency. Libvirt builds
without it just fine, but then fails at runtime:

  error: failed to connect to the hypervisor
  error: internal error: unable to initialize VirtualBox driver API

This is because VBoxXPCOMC.so is loaded at runtime (via dlopen()
[1]). But all of our ebuilds are missing the dependency because
of the commit below (which has really poor commit message and
doesn't explain why the dependency was removed).

For now, libvirt supports virtualbox-6.1 and virtualbox-7.0.
Support for virtualbox-7.1 is being worked on [2].

This reverts commit cbb7c820c7c34033be2bff54585804559145985a.

1: https://gitlab.com/libvirt/libvirt/-/blob/master/src/vbox/vbox_XPCOMCGlue.c?ref_type=heads#L110
2: https://gitlab.com/libvirt/libvirt/-/issues/681
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>