media-gfx/imagemagick: workaround GCC 13 bug
-O3 implies -funswitch-loops which triggers an unfortunate rare edgecase
with one of the switch statements in ImageMagick. Apply a workaround
for GCC 13 for now (14+ is fixed already and it's latent otherwise).

The fix will land on the 13 branch shortly upstream but let's get the fix
out quickly on our end so we don't have to rush to stable it and so on
(although we will of course do it soon).

Bug: https://github.com/ImageMagick/ImageMagick/issues/7677
Bug: https://gcc.gnu.org/PR117100
Closes: https://bugs.gentoo.org/941208
Signed-off-by: Sam James <sam@gentoo.org>