www-client/chromium: add 131.0.6753.0
This commit brings the development channel into ::gentoo.
We may remove it if the overheads prove too much, and
it is not a priority for updates.

The hope is that the development slot, if kept up-to-date,
will provide additional time to adapt the ebuild to upcoming
major versions of Chromium before they hit stable and become
security issues.

Changes:
  - use `filter-lto` and rely on GN to appropriately set LTO options.
  - Remove `updater` force-disable - it shouldn't be enabled anyway.

Signed-off-by: Matt Jolly <kangie@gentoo.org>