www-client/chromium: Enable Control-flow Integrity (CFI)
This relies on Clang's CFI mechanisms, and is dependent on
split LTO Units, which require a "nightly" Rust, years after
the feature's implementation, with it being in widespread use
in the real world for software packaged on stable systems.

Because there is no better way to go about this we are forced
to YOLO it and set `RUSTC_BOOTSTRAP=1` in the ebuild. Google's
default build uses a nightly compiler, so this should not cause
any issues.

CFI will be enabled on a platform if LTO is enabled, unless this
needs to be filtered for non x86_64 systems.

According to Google this should result in a slight (<1%) overhead
in terms of CPU and a small increase (<10%) to the size
of the compiled binary. Memory overhead is minimal.

Signed-off-by: Matt Jolly <kangie@gentoo.org>